This is the now latest whole thing
NetBeans project.
This is the now latest whole thing. Please igonre any previous upload because there is a huge diference with the previous uploads.
If you still working on previous one, it will be an extreme pain for me to add it to the new version since the architecture is not the same.